style changes and minor fixes

+++ bike-rental/routes/admin.js
@@ -12,8 +12,8 @@ const adminController = require('../controllers/adminController');
 
   router.post('/', async (req, res, next) => {
     try {
-      const bikeType = Number.parseInt(req.body.bikeType); 
-      const dailyRate = Number.parseInt(req.body.dailyRate);
+      const bikeType = Number.parseFloat(req.body.bikeType); 
+      const dailyRate = Number.parseFloat(req.body.dailyRate);
       const statusId = 1;
